move AboutFragment to another package
